Revert "Move Form Builder to ee-components"
This reverts commit 2b0154549b67d435e69c5fc16a4ae8ed4bca3a6a.
