+++ packages/ui-components/src/FormBuilder/FormBuilder.tsx
@@ -1,8 +1,7 @@
 import { useMemo } from 'react';
 import classNames from 'classnames';
 
-import { Container } from '@eventespresso/ui-components';
-
+import { Container } from '../Container';
 import { FormBuilderSidebar } from './FormBuilderSidebar';
 import { FormSection } from './FormSection';
 import { useOpenElement } from './useOpenElement';
@@ -35,8 +34,8 @@ const FormBuilder: React.FC<FormBuilderProps> = ({
 	// controls and tracks which element is open for editing
 	const { isOpen, toggleElement } = useOpenElement();
 
-	const form = formSections.map((formSection) => (
-		<FormSection key={formSection.UUID} formSection={formSection} isOpen={isOpen} toggleElement={toggleElement} />
+	const form = formSections.map((formSection, index) => (
+		<FormSection key={index} formSection={formSection} isOpen={isOpen} toggleElement={toggleElement} />
 	));
 
 	return (
