+++ packages/ui-components/src/FormBuilder/FormElement/FormElement.tsx
@@ -1,8 +1,7 @@
 import classNames from 'classnames';
-
 import { More } from '@eventespresso/icons';
-import { IconButton } from '@eventespresso/ui-components';
 
+import { IconButton } from '../../Button';
 import { FormElementInput } from './FormElementInput';
 import { FormElementToolbar } from './FormElementToolbar';
 import { FormElementTabs } from './Tabs';
