+++ packages/ui-components/src/FormBuilder/FormElement/FormElementToolbar.tsx
@@ -1,8 +1,7 @@
 import classNames from 'classnames';
-
 import { DragHandle, Trash } from '@eventespresso/icons';
-import { IconButton } from '@eventespresso/ui-components';
 
+import { IconButton } from '../../Button';
 import { ELEMENT_BLOCKS_INDEXED } from '../constants';
 
 import type { FormElementToolbarProps } from '../types';
