+++ packages/ui-components/src/FormBuilder/FormElement/MappedElement.tsx
@@ -2,8 +2,7 @@ import { pick } from 'ramda';
 import { useMemo } from 'react';
 
 import type { AnyObject } from '@eventespresso/utils';
-import { TextInput, NumberInput, Textarea, Select, Switch, withLabel } from '@eventespresso/ui-components';
-
+import { TextInput, NumberInput, Textarea, Select, Switch, withLabel } from '../../';
 import { FormInputProps, FormInput } from '../types';
 
 const DefaultComponent = () => null;
