+++ packages/ui-components/src/FormBuilder/FormElement/Tabs/Settings.tsx
@@ -1,6 +1,7 @@
 import { __ } from '@eventespresso/i18n';
-import { TextInput, withLabel } from '@eventespresso/ui-components';
 
+import { TextInput } from '../../../text-input';
+import { withLabel } from '../../../withLabel';
 import type { SettingsProps } from '../../types';
 
 const TextWithLabel = withLabel(TextInput);
