+++ packages/ui-components/src/index.ts
@@ -30,6 +30,7 @@ export * from './EntityList';
 export * from './EntityPaperFrame';
 export * from './EspressoTable';
 export * from './FilterBarFilter';
+export * from './FormBuilder';
 export * from './Grid';
 export * from './Heading';
 export * from './HelpIcon';
